+++ source/extensions/filters/network/echo/echo.h
@@ -6,13 +6,13 @@
 
 namespace Envoy {
 namespace Extensions {
-namespace NetworkFilter {
+namespace NetworkFilters {
 namespace Echo {
 
 /**
  * Implementation of a basic echo filter.
  */
-class Instance : public Network::ReadFilter, Logger::Loggable<Logger::Id::filter> {
+class EchoFilter : public Network::ReadFilter, Logger::Loggable<Logger::Id::filter> {
 public:
   // Network::ReadFilter
   Network::FilterStatus onData(Buffer::Instance& data, bool end_stream) override;
@@ -26,6 +26,6 @@ private:
 };
 
 } // namespace Echo
-} // namespace NetworkFilter
+} // namespace NetworkFilters
 } // namespace Extensions
 } // namespace Envoy
