+++ test/integration/echo_integration_test.cc
@@ -3,7 +3,6 @@
 #include "test/server/utility.h"
 
 namespace Envoy {
-namespace {
 
 std::string echo_config;
 
@@ -60,8 +59,6 @@ TEST_P(EchoIntegrationTest, Hello) {
   EXPECT_EQ("hello", response);
 }
 
-// TODO(mattklein123): This test does not belong here in an extension integration test. Move it
-//                     somewhere else.
 TEST_P(EchoIntegrationTest, AddRemoveListener) {
   const std::string json = TestEnvironment::substitute(R"EOF(
   {
@@ -125,5 +122,4 @@ TEST_P(EchoIntegrationTest, AddRemoveListener) {
   connection2.run();
 }
 
-} // namespace
 } // namespace Envoy
