--- test/integration/extensions/filters/network/echo/BUILD
@@ -1,21 +1,0 @@
-licenses(["notice"])  # Apache 2
-
-load(
-    "//bazel:envoy_build_system.bzl",
-    "envoy_cc_test",
-    "envoy_package",
-)
-
-envoy_package()
-
-envoy_cc_test(
-    name = "echo_integration_test",
-    srcs = [
-        "echo_integration_test.cc",
-    ],
-    deps = [
-        "//source/extensions/filters/network/echo:echo_lib",
-        "//test/integration:integration_lib",
-        "//test/server:utility_lib",
-    ],
-)
